requirements.txt
asgiref
boto3
botocore
Django
django-cors-headers
django-storages
djangorestframework
djangorestframework_simplejwt
jmespath
pillow
psycopg2-binary
PyJWT
python-dateutil
python-dotenv
s3transfer
six
sqlparse
typing_extensions
tzdata
urllib3
